修改样式问题

+++ src/main/java/com/isoler/studynav/business/fl/model/bean/Fl.java
@@ -46,7 +46,7 @@ public class Fl extends BaseEntity {
 
     @ApiModelProperty(value = "排序")
     @TableField("n_order")
-    private Integer xh;
+    private Long xh;
 
     @ApiModelProperty(value = "网址列表")
     @TableField(exist = false)
