+++ src/main/java/com/isoler/studynav/business/link/model/bean/Link.java
@@ -55,7 +55,7 @@ public class Link extends BaseEntity {
 
     @ApiModelProperty(value = "网址排序")
     @TableField("n_order")
-    private Integer xh;
+    private Long xh;
 
     @ApiModelProperty(value = "分类名称")
     @TableField(exist = false)
